# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data Modelling API",
    "description": "REST API for data modeling, schema management, and collaboration",
    "contact": {
      "name": "API Support",
      "email": "mark@opendatamodelling.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.1.1"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://api.opendatamodelling.com/api/v1",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/collaboration/sessions/{session_id}": {
      "get": {
        "tags": [
          "Collaboration"
        ],
        "summary": "GET /collaboration/sessions/{session_id} - Get a collaboration session",
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Collaboration"
        ],
        "summary": "DELETE /collaboration/sessions/{session_id} - End a collaboration session",
        "operationId": "end_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session ended successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}